+++ src/client/src/api-lib/index.js
@@ -1,7 +1,6 @@
 import { getHeight } from "./actions";
 import { deleteBoreholes, patchBoreholes } from "./actions/borehole";
 import { loadSettings, patchSettings } from "./actions/settings";
-import { draftTerms, getTerms, getTermsDraft, publishTerms } from "./actions/terms";
 import { setAuthentication, unsetAuthentication } from "./actions/user";
 import store, { injectReducer } from "./reducers";
 
@@ -9,10 +8,6 @@ export {
   getHeight,
   loadSettings,
   patchSettings,
-  draftTerms,
-  getTerms,
-  getTermsDraft,
-  publishTerms,
   setAuthentication,
   unsetAuthentication,
   deleteBoreholes,